+++ plugins_init.h
@@ -23,11 +23,6 @@
   along with Grbl.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#if PLASMA_ENABLE
-    extern void plasma_init (void);
-    plasma_init();
-#endif
-
 #if TRINAMIC_ENABLE
     extern bool trinamic_init (void);
     trinamic_init();